requirements.txt
pillow
customtkinter
requests
SpeechRecognition
pydub